requirements.txt
aiosqlite==0.18.0
alembic==1.9.1
anyio==3.6.2
appnope==0.1.3
apprise==1.2.0
asgi-lifespan==2.0.0
asttokens==2.2.1
asyncpg==0.27.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
coolname==2.1.0
croniter==1.3.8
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
docker==6.0.1
entrypoints==0.4
executing==1.2.0
fastapi==0.88.0
Flask==2.2.2
Flask-Caching==2.0.1
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
fsspec==2022.11.0
google-auth==2.15.0
greenlet==2.0.1
griffe==0.25.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==5.2.0
ipykernel==6.19.4
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jupyter_client==7.4.8
jupyter_core==5.1.1
kiwisolver==1.4.4
kubernetes==25.3.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
oauthlib==3.2.2
orjson==3.8.3
packaging==22.0
pandas==1.5.2
parso==0.8.3
pathspec==0.10.3
patsy==0.5.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
prefect==2.7.4
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
pytz==2022.7
pytzdata==2020.1
PyYAML==6.0
pyzmq==24.0.1
readchar==4.0.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
stack-data==0.6.2
starlette==0.22.0
statsmodels==0.13.5
text-unidecode==1.3
toml==0.10.2
tornado==6.2
traitlets==5.8.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
wcwidth==0.2.5
websocket-client==1.4.2
Werkzeug==2.2.2
zipp==3.11.0
